+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+## gDRplots 0.0.12 - 2024-07-12
+* move `get_combo_col_settings` and `get_iso_colors` from `gDRutils` package
+
 ## gDRplots 0.0.11 - 2024-07-11
 * add plots to quality control for combo data - combo metrics and reference isoline